+++ origami/collections/src/app-elements/app-elements.ts
@@ -1,0 +1,9 @@
+import { Directive } from '@angular/core';
+import { EmitChangesDirective } from '@codebakery/origami';
+
+/* @private */
+@Directive({
+  selector: `app-box, app-drawer, app-drawer-layout, app-header, app-header-layout,
+    app-scrollpos-control, app-toolbar, app-localize-behavior, app-pouchdb, app-route, app-storage`
+})
+export class AppElement extends EmitChangesDirective { }