+++ origami/collections/src/collections.ts
@@ -1,0 +1,6 @@
+export * from './app-elements/index';
+export * from './gold-elements/index';
+export * from './iron-elements/index';
+export * from './molecules/index';
+export * from './paper-elements/index';
+export * from './platinum-elements/index';