+++ origami/collections/src/paper-elements/paper-elements.ts
@@ -1,0 +1,7 @@
+import { Directive } from '@angular/core';
+import { EmitChangesDirective } from '@codebakery/origami';
+
+@Directive({
+  selector: `paper-ripple`
+})
+export class PaperElement extends EmitChangesDirective { }