+++ origami/src/util/customElements.ts
@@ -1,0 +1,3 @@
+export function getCustomElements(): CustomElementRegistry {
+  return window.customElements;
+}