+++ src/collections/iron-elements/iron-input-elements.ts
@@ -1,8 +1,0 @@
-import { Directive } from '@angular/core';
-
-import { EmitChangesDirective } from '../../events/emit-changes.directive';
-
-@Directive({
-  selector: `iron-dropdown, iron-form, iron-input, iron-label`
-})
-export class IronInputElement extends EmitChangesDirective { }