+++ src/collections/paper-elements/paper-elements.ts
@@ -1,8 +1,0 @@
-import { Directive } from '@angular/core';
-
-import { EmitChangesDirective } from '../../events/emit-changes.directive';
-
-@Directive({
-  selector: `paper-ripple`
-})
-export class PaperElement extends EmitChangesDirective { }